general: refactor how update jobs are scheduled
Update jobs used to be long running async tasks, and are signaled using
an event channel whenever the prompt needs to be refreshed. This has a
problem however, when a new round updates is started before the previous
round is finished, it become impossible to track how many tasks are
actually running. This also means the state file could get updated with
stale information from the previous round. What we really want to do is
to cancel all the previous updates, close update message channels, and
start over.

So with this commit, update tasks are short running and will resolve
once update has finished. Whenever a refresh is desired, all current
running tasks are dropped and thus cancelled.

This commit also solves another race condition. It's possible that the
server sends the shell a SIGUSR1 when the shell is in the middle of a
prompt repaint, in which case another repaint won't be started. If this
SIGUSR1 is the last SIGUSR1 the server sends, the prompt will stuck in
an incomplete state. This problem is already solved by periodically
triggering repaint until the renderer sees a completed state file and
send the server a SIGUSR2. But it's impossible for the server to know
if the completed state file the renderer saw was the latest one. The
solution is including a version number in the state file, and having the
renderer include this version number to the server in the SIGUSR2 it
sends in siginfo_t. Since no existing Rust signal handling crate exposes
siginfo_t, I forked async-signal and added my own.